\name{EMMA.GUI}
\alias{EMMA.GUI}
\title{Function to run EMMA with a GUI.}
\description{This function will call a shiny app that allows running EMMA with 
  a graphic user interface.}
\usage{EMMA.GUI(path = "~/Documents/projects/R-packages/R_EMMAgeo/EMMAgeo/")}
\arguments{
  \item{path}{Character scalar, path to the R-package, with package name. If 
     omitted, the likely path will be searched automatically.}
}

\value{A graphic user interface to run EMMA.}
\author{Michael Dietze, Elisabeth Dietze}
\seealso{\code{\link{EMMA}}}
\examples{
## not run
# EMMA.GUI()
}

\keyword{EMMA}
