\name{EMMAgeo-package}
\alias{EMMAgeo}
\docType{package}
\title{End-member modelling algorithm and supporting functions for grain-size analysis}
\description{This package provides a set of functions for convenient end-member modelling analysis of grain-size data (EMMAgeo).}
\details{
\tabular{ll}{ Package: \tab EMMAgeo\cr Type: \tab Package\cr Version:
\tab 0.9.2\cr Date: \tab 2015-06-07\cr License: \tab GPL-3\cr }
}
\author{Michael Dietze, Elisabeth Dietze}
\keyword{package}
