% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMAgeo-package.R
\docType{data}
\name{EMpot}
\alias{EMpot}
\title{example data}
\format{The format is: List of 8 $ q : num [1:90] 4 4 4 4 4 4 4 4 4 4 ...  $
lw : num [1:90] 0 0 0 0 0.05 0.05 0.05 0.05 0.1 0.1 ...  $ modes : num
[1:90] 12 32 61 80 12 32 61 80 12 32 ...}
\description{
A list with output of the function test.robustness()
}
\details{
The dataset is the result of the function test.robustness() of the R-package
EMMAgeo.
}
\examples{

## load example data set
data(example_EMpot)

}
\keyword{datasets}
