% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMAgeo-package.R
\docType{data}
\name{EMrob}
\alias{EMrob}
\title{example data}
\format{The format is: List of 12 $ Vqsn.data :List of 4 ..$ : num [1:15,
1:80] 0.18929 0.184 0.18304 0.00698 0.02033 ...}
\description{
Robust end-members, a list with output of the function robust.EM()
}
\details{
The dataset is the result of the function robust.EM() of the R-package
EMMAgeo.
}
\examples{

## load example data set
data(example_EMrob)

}
\keyword{datasets}

