\name{REM}
\alias{REM}
\docType{data}
\title{example data}
\description{Robust end-members, a list with output of the function robust.EM()}
\usage{data(REM)}
\format{
  The format is:
 List of 12
 $ Vqsn.data  :List of 4
  ..$ : num [1:15, 1:80] 0.18929 0.184 0.18304 0.00698 0.02033 ...
}
\details{The dataset is the result of the function robust.EM() of the R-package
         EMMAgeo.}
\examples{
## load example data set
data(examples)
}
\keyword{datasets}
