\name{SEM.data}
\alias{SEM.data}
\docType{data}
\title{example data}
\description{Depth and diameter of SEM-imaged grains}
\usage{data(SEM.data)}
\format{
    The format is:
  num [1:1237, 1:2] 5.45 2.59 2.77 4.43 4.85 ...
}
\details{The dataset is the result of a classified and digitised SEM image
         and contains two columns: depth and diameter, both in micrometers.}
\examples{
## load example data set
data(examples)
}
\keyword{datasets}
