% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/server.R
\name{shinyServer}
\alias{shinyServer}
\title{Function to generate GUI server for EMMA.}
\usage{
shinyServer(input, output)
}
\arguments{
\item{input}{A collection of input objects.}

\item{output}{A collection of input objects.}
}
\value{
A GUI server.
}
\description{
A graphical user interface (GUI) server backend is started.
}
\examples{
## Not run
# EMMA.GUI()
}
\author{
Michael Dietze, Elisabeth Dietze
}
\keyword{EMMA}

