\name{TR}
\alias{TR}
\docType{data}
\title{example data}
\description{A list with output of the function test.robustness()}
\usage{data(TR)}
\format{
  The format is:
 List of 8
 $ q       : num [1:90] 4 4 4 4 4 4 4 4 4 4 ...
 $ lw      : num [1:90] 0 0 0 0 0.05 0.05 0.05 0.05 0.1 0.1 ...
 $ modes   : num [1:90] 12 32 61 80 12 32 61 80 12 32 ...
}
\details{The dataset is the result of the function test.robustness() of 
         the R-package EMMAgeo.}
\examples{
## load example data set
data(examples)
}
\keyword{datasets}
