% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMMAgeo-package.R
\docType{data}
\name{X.artificial}
\alias{X.artificial}
\title{example data}
\format{The format is: num [1:80, 1:80] 0.00032 0.00057 0.00037 0.00029 ...}
\description{
Artificial data set created by randomly mixed normal distributions
}
\details{
The dataset is the result of five randomly mixed normal distributions,
forming 80 samples, each represented by 80 classes. Some white noise is
added.
}
\examples{
## load example data set
data(X.artificial)

## plot first 10 samples stacked in one line plot
plot(NA, xlim = c(1, 80), ylim = c(1, 14))
for(i in 1:10) {lines(1:80, X.artificial[i,] * 50 + i)}
}
\keyword{datasets}

