% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMAgeo-package.R
\docType{data}
\name{X}
\alias{X}
\title{example data}
\format{
num [1:100, 1:116] 0.000899 0.000516 0.00136 0.000989 0.00102 ...
}
\description{
Synthetic data set created by randomly mixed natural end-members
}
\details{
The dataset is the result of four mixed natural end-members.
}
\examples{

## load example data set
data(example_X)

## extract grain-size classes
s <- as.numeric(colnames(X))

## plot first 10 samples stacked in one line plot
plot(NA, 
     xlim = c(1, ncol(X)), 
     ylim = c(1, 20))
     
for(i in 1:10) {
  lines(x = s, 
        y = X[i,] + i)
}

## plot grain-size map
image(x = s, 
      z = t(X), 
      log = "x", 
      col = rainbow(n = 250))

}
\keyword{datasets}
