\name{X.to.find.EMMAgeo.path.MDT}
\alias{X.to.find.EMMAgeo.path.MDT}
\docType{data}
\title{example data}
\description{This file is simply used to securely locate the package path}
\usage{data(X.to.find.EMMAgeo.path.MDT)}
\format{
  The format is:
 List of 12
 $ Vqsn.data  :List of 4
  ..$ : num [1:15, 1:80] 0.18929 0.184 0.18304 0.00698 0.02033 ...
}
\details{No details needed}
\examples{
## load example data set
data(examples)
}
\keyword{datasets}