\name{EMMAgeo-package}
\alias{EMMAgeo-package}
\alias{EMMAgeo}
\docType{package}
\title{End-member modelling algorithm and supporting functions for grain-size analysis}
\description{This package provides a set of functions for convenient end-member modelling analysis of grain-size data (EMMAgeo).}
\details{
\tabular{ll}{Package: \tab EMMAgeo\cr
Type: \tab Package\cr
Title: \tab End-member modelling algorithm and supporting functions for grain-size analysis\cr
Version: \tab 0.9.1\cr
Date: \tab 2013-10-14\cr
Author: \tab Michael Dietze, Elisabeth Dietze\cr
Maintainer: \tab Michael Dietze <micha.dietze@mailbox.tu-dresden.de>\cr
License: \tab GPL-3\cr
Depends: \tab R (>= 3.0.1), GPArotation, limSolve, shape\cr}
}
\author{Michael Dietze, Elisabeth Dietze}

\keyword{ package }


