% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMMAgeo-package.R
\docType{package}
\name{EMMAgeo-package}
\alias{EMMAgeo}
\alias{EMMAgeo-package}
\title{End-member modelling algorithm and supporting functions for grain-size
analysis}
\description{
This package provides a set of functions for end-member modelling
analysis of grain-size data (EMMAgeo).
}
\details{
\tabular{ll}{ Package: \tab EMMAgeo\cr Type: \tab Package\cr Version: \tab
0.9.2\cr Date: \tab 2015-06-07\cr License: \tab GPL-3\cr }
}
\author{
Michael Dietze, Elisabeth Dietze
}
\keyword{package}

