% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R
\name{GUI}
\alias{GUI}
\title{Start GUI for EMMA}
\usage{
GUI(...)
}
\arguments{
\item{...}{further arguments to pass to \code{\link{runApp}}}
}
\description{
This function starts a browser-based graphic user interface for EMMA. The 
GUI has so far been tested on a Linux system, both with the browser of 
RStudio and Mozilla Firefox. It permits basic access to import, display 
and model a user-provided data set.
}
\examples{

\dontrun{
# Start the GUI
GUI()
}

}
\author{
Michael Dietze
}
\seealso{
\code{\link{runApp}}
}

