% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GUI.R
\name{GUI}
\alias{GUI}
\title{Start GUI for EMMA}
\usage{
GUI(...)
}
\arguments{
\item{...}{further arguments to pass to \code{\link{runApp}}}
}
\description{
This function starts a browser-based graphic user interface for EMMA.
}
\examples{
\dontrun{
# Start the GUI
GUI()
}
}
\author{
Michael Dietze
}
\seealso{
\code{\link{runApp}}
}

