% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMMAgeo-package.R
\docType{data}
\name{SEM.data}
\alias{SEM.data}
\title{example data}
\format{The format is: num [1:1237, 1:2] 5.45 2.59 2.77 4.43 4.85 ...}
\description{
Depth and diameter of SEM-imaged grains
}
\details{
The dataset is the result of a classified and digitised SEM image and
contains two columns: depth and diameter, both in micrometers.
}
\examples{
## load example data set
data(examples)
}
\keyword{datasets}

