% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMMAgeo-package.R
\docType{data}
\name{X.to.find.EMMAgeo.path.MDT}
\alias{X.to.find.EMMAgeo.path.MDT}
\title{example data}
\format{The format is: List of 12 $ Vqsn.data :List of 4 ..$ : num [1:15,
1:80] 0.18929 0.184 0.18304 0.00698 0.02033 ...}
\description{
This file is simply used to securely locate the package path
}
\details{
No details needed
}
\examples{
## load example data set
data(examples)
}
\keyword{datasets}

